added unit tests
added exceptions

updated style code

minor fixes

+++ service/src/main/java/com/epam/esm/service/impl/GiftCertificateServiceImpl.java
@@ -9,8 +9,11 @@ import com.epam.esm.repository.GiftCertificateRepository;
 import com.epam.esm.repository.TagRepository;
 import com.epam.esm.repository.exception.RepositoryException;
 import com.epam.esm.service.GiftCertificateService;
-import com.epam.esm.service.exception.GiftCertificateNotFoundException;
+import com.epam.esm.service.exception.certificate.GiftCertificateAlreadyExistException;
+import com.epam.esm.service.exception.certificate.GiftCertificateNotFoundException;
 import com.epam.esm.service.exception.ServiceException;
+import com.epam.esm.service.exception.certificate.UnableDeleteGiftCertificateException;
+import com.epam.esm.service.exception.certificate.UnableUpdateGiftCertificate;
 import com.epam.esm.util.ParamsUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -19,6 +22,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -36,29 +40,32 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     @Override
     public List<GiftCertificateDto> findAllWithTags(Map<String, String[]> params) throws ServiceException {
         try {
-
             // TODO: 13-Jan-21 validate params
-
             if (ParamsUtil.hasParams(ParamsUtil.SORT_PARAM, params)) {
                 Sort sort = ParamsUtil.buildSortByParams(ParamsUtil.getSortParams(params));
                 return findAllWithTagsByOrder(params, sort);
             }
 
-            if (ParamsUtil.hasParams(ParamsUtil.TAG_PARAM, params) && ParamsUtil.hasParams(ParamsUtil.PART_PARAM, params)) {
+            if (ParamsUtil.hasParams(ParamsUtil.TAG_PARAM, params)
+                    && ParamsUtil.hasParams(ParamsUtil.PART_PARAM, params)) {
                 Set<GiftCertificate> giftCertificateSet =
-                        giftCertificateRepository.findAllByTagNameAndPartOfNameOrDescription(ParamsUtil.getTagParam(params), ParamsUtil.getPartParam(params));
+                        giftCertificateRepository.findAllByTagNameAndPartOfNameOrDescription(
+                                ParamsUtil.getTagParam(params), ParamsUtil.getPartParam(params));
                 return giftCertificateConverter.entityToDtoList(new ArrayList<>(giftCertificateSet));
             }
             if (ParamsUtil.hasParams(ParamsUtil.TAG_PARAM, params)) {
-                Set<GiftCertificate> giftCertificateSet = giftCertificateRepository.findAllByTagName(ParamsUtil.getTagParam(params));
+                Set<GiftCertificate> giftCertificateSet =
+                        giftCertificateRepository.findAllByTagName(ParamsUtil.getTagParam(params));
                 return giftCertificateConverter.entityToDtoList(new ArrayList<>(giftCertificateSet));
             }
             if (ParamsUtil.hasParams(ParamsUtil.PART_PARAM, params)) {
-                Set<GiftCertificate> giftCertificateSet = giftCertificateRepository.findAllByPartOfNameOrDescription(ParamsUtil.getPartParam(params));
+                Set<GiftCertificate> giftCertificateSet =
+                        giftCertificateRepository.findAllByPartOfNameOrDescription(ParamsUtil.getPartParam(params));
                 return giftCertificateConverter.entityToDtoList(new ArrayList<>(giftCertificateSet));
             }
 
             Set<GiftCertificate> giftCertificateSet = giftCertificateRepository.findAllWithTags();
+
             return giftCertificateConverter.entityToDtoList(new ArrayList<>(giftCertificateSet));
         } catch (RepositoryException e) {
             throw new ServiceException(e);
@@ -67,24 +74,29 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     private List<GiftCertificateDto> findAllWithTagsByOrder(Map<String, String[]> params, Sort sort) throws ServiceException {
         try {
-
             // TODO: 14-Jan-21 validate sort, params
-
-            if (ParamsUtil.hasParams(ParamsUtil.TAG_PARAM, params) && ParamsUtil.hasParams(ParamsUtil.PART_PARAM, params)) {
+            if (ParamsUtil.hasParams(ParamsUtil.TAG_PARAM, params)
+                    && ParamsUtil.hasParams(ParamsUtil.PART_PARAM, params)) {
                 Set<GiftCertificate> giftCertificateSet =
-                        giftCertificateRepository.findAllByTagNameAndPartOfNameOrDescriptionByOrder(ParamsUtil.getTagParam(params), ParamsUtil.getPartParam(params), sort);
+                        giftCertificateRepository.findAllByTagNameAndPartOfNameOrDescriptionByOrder(
+                                ParamsUtil.getTagParam(params), ParamsUtil.getPartParam(params), sort);
                 return giftCertificateConverter.entityToDtoList(new ArrayList<>(giftCertificateSet));
             }
             if (ParamsUtil.hasParams(ParamsUtil.TAG_PARAM, params)) {
-                Set<GiftCertificate> giftCertificateSet = giftCertificateRepository.findAllByTagNameByOrder(ParamsUtil.getTagParam(params), sort);
+                Set<GiftCertificate> giftCertificateSet =
+                        giftCertificateRepository.findAllByTagNameByOrder(
+                                ParamsUtil.getTagParam(params), sort);
                 return giftCertificateConverter.entityToDtoList(new ArrayList<>(giftCertificateSet));
             }
             if (ParamsUtil.hasParams(ParamsUtil.PART_PARAM, params)) {
-                Set<GiftCertificate> giftCertificateSet = giftCertificateRepository.findAllByPartOfNameOrDescriptionByOrder(ParamsUtil.getPartParam(params), sort);
+                Set<GiftCertificate> giftCertificateSet =
+                        giftCertificateRepository.findAllByPartOfNameOrDescriptionByOrder(
+                                ParamsUtil.getPartParam(params), sort);
                 return giftCertificateConverter.entityToDtoList(new ArrayList<>(giftCertificateSet));
             }
 
             Set<GiftCertificate> giftCertificateSet = giftCertificateRepository.findAllWithTagsByOrder(sort);
+
             return giftCertificateConverter.entityToDtoList(new ArrayList<>(giftCertificateSet));
         } catch (RepositoryException e) {
             throw new ServiceException(e);
@@ -104,6 +116,7 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     @Override
     public GiftCertificateDto findByName(String name) throws ServiceException {
         try {
+            // TODO: 16-Jan-21 validate name
             GiftCertificate giftCertificate = giftCertificateRepository.findByName(name);
 
             if (giftCertificate == null) {
@@ -130,7 +143,12 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     @Override
     public GiftCertificateDto save(GiftCertificateDto giftCertificateDto) throws ServiceException {
         try {
+            // TODO: 16-Jan-21 validate giftCertificateDto
             GiftCertificate giftCertificate = giftCertificateConverter.dtoToEntity(giftCertificateDto);
+
+            if (giftCertificateRepository.findByName(giftCertificate.getName()) != null) {
+                throw new GiftCertificateAlreadyExistException(ServiceError.GIFT_CERTIFICATE_ALREADY_EXISTS.getCode());
+            }
 
             if (giftCertificate.getTags() != null) {
                 Set<Tag> tags = giftCertificate.getTags().stream()
@@ -141,6 +159,7 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
                 giftCertificate.setTags(tags);
             }
 
+            giftCertificate.setCreateDate(LocalDateTime.now());
             giftCertificate = giftCertificateRepository.save(giftCertificate);
 
             return giftCertificateConverter.entityToDto(giftCertificate);
@@ -153,13 +172,14 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     @Override
     public GiftCertificateDto updateByName(String name, GiftCertificateDto giftCertificateDto) throws ServiceException {
         try {
+            // TODO: 16-Jan-21 validate name, giftCertificateDto
             GiftCertificate giftCertificate = giftCertificateRepository.findByName(name);
 
             if (giftCertificate == null) {
-                throw new GiftCertificateNotFoundException("Gift Certificate not found");
+                throw new GiftCertificateNotFoundException(ServiceError.GIFT_CERTIFICATE_NOT_FOUNT.getCode());
             }
 
-            final Integer id = giftCertificate.getId();
+            final Long id = giftCertificate.getId();
             giftCertificate = giftCertificateConverter.dtoToEntity(giftCertificateDto);
 
             Set<Tag> newTags = new HashSet<>();
@@ -171,7 +191,12 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
                 giftCertificate.setTags(newTags);
             }
 
-            giftCertificateRepository.updateById(id, giftCertificate);
+            giftCertificate.setLastUpdateDate(LocalDateTime.now());
+            final long count = giftCertificateRepository.updateById(id, giftCertificate);
+
+            if (count == 0) {
+                throw new UnableUpdateGiftCertificate(ServiceError.UNABLE_UPDATE_GIFT_CERTIFICATE.getCode());
+            }
 
             giftCertificate = giftCertificateRepository.findById(id);
             giftCertificate.setTags(newTags);
@@ -182,10 +207,22 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
         }
     }
 
+    @Transactional
     @Override
     public void deleteByName(String name) throws ServiceException {
         try {
-            giftCertificateRepository.deleteByName(name);
+            // TODO: 16-Jan-21 validate name
+            GiftCertificate giftCertificate = giftCertificateRepository.findByName(name);
+
+            if (giftCertificate == null) {
+                throw new GiftCertificateNotFoundException(ServiceError.GIFT_CERTIFICATE_NOT_FOUNT.getCode());
+            }
+
+            final long count = giftCertificateRepository.deleteByName(name);
+
+            if (count == 0) {
+                throw new UnableDeleteGiftCertificateException(ServiceError.GIFT_CERTIFICATE_UNABLE_DELETE.getCode());
+            }
         } catch (RepositoryException e) {
             throw new ServiceException(e);
         }
