+++ service/src/main/java/com/epam/esm/service/impl/TagServiceImpl.java
@@ -8,11 +8,14 @@ import com.epam.esm.repository.exception.RepositoryException;
 import com.epam.esm.service.TagService;
 import com.epam.esm.dto.TagDto;
 import com.epam.esm.service.exception.ServiceException;
-import com.epam.esm.service.exception.TagNotFoundException;
+import com.epam.esm.service.exception.tag.TagAlreadyExistException;
+import com.epam.esm.service.exception.tag.TagNotFoundException;
+import com.epam.esm.service.exception.tag.UnableDeleteTagException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -28,8 +31,8 @@ public class TagServiceImpl implements TagService {
     @Override
     public List<TagDto> findAll() throws ServiceException {
         try {
-            List<Tag> tagList = tagRepository.findAll();
-            return tagConverter.entityToDtoList(tagList);
+            List<Tag> tags = tagRepository.findAll();
+            return tagConverter.entityToDtoList(tags);
         } catch (RepositoryException e) {
             throw new ServiceException(e);
         }
@@ -38,6 +41,7 @@ public class TagServiceImpl implements TagService {
     @Override
     public TagDto findByName(String name) throws ServiceException {
         try {
+            // TODO: 15-Jan-21 validate name
             Tag tag = tagRepository.findByName(name);
             if (tag == null) {
                 throw new TagNotFoundException(ServiceError.TAG_NOT_FOUND.getCode());
@@ -51,18 +55,33 @@ public class TagServiceImpl implements TagService {
     @Override
     public TagDto save(TagDto tagDto) throws ServiceException {
         try {
+            // TODO: 15-Jan-21 validate tagDto
             Tag tag = tagConverter.dtoToEntity(tagDto);
-            tag = tagRepository.save(tag);
-            return tagConverter.entityToDto(tag);
+            if (tagRepository.findByName(tag.getName()) != null) {
+                throw new TagAlreadyExistException(ServiceError.TAG_ALREADY_EXISTS.getCode());
+            }
+            return tagConverter.entityToDto(tagRepository.save(tag));
         } catch (RepositoryException e) {
             throw new ServiceException(e);
         }
     }
 
+    @Transactional
     @Override
     public void deleteByName(String name) throws ServiceException {
         try {
-            tagRepository.deleteByName(name);
+            // TODO: 16-Jan-21 validate name
+            Tag tag = tagRepository.findByName(name);
+
+            if (tag == null) {
+                throw new TagNotFoundException(ServiceError.TAG_NOT_FOUND.getCode());
+            }
+
+            long count = tagRepository.deleteByName(name);
+
+            if (count == 0) {
+                throw new UnableDeleteTagException(ServiceError.TAG_UNABLE_DELETE.getCode());
+            }
         } catch (RepositoryException e) {
             throw new ServiceException(e);
         }
